requirements.txt
# Core Dependencies
langchain==0.1.0
langchain-google-genai==0.0.6
google-generativeai==0.3.2
python-dotenv==1.0.0
transformers

# UI Dependencies
streamlit==1.30.0
streamlit-chat==0.1.1

# Data Processing
pandas==2.1.4
numpy==1.26.3

# Utilities
pydantic==2.5.3
typing-extensions==4.9.0